+++ Packages/cat.hisa.asset-patcher/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Create output directory if not exists during patch.
 - Default language loaded from system if it supports.
 
 ## [1.0.0] - 2024-04-10